+++ src/components/ui/forms/input-otp.tsx
@@ -39,7 +39,7 @@ const InputOTPSlot = React.forwardRef<
 >(({ index, className, ...props }, ref) => {
   const inputOTPContext = React.useContext(OTPInputContext)
   // eslint-disable-next-line security/detect-object-injection
-  const { char, hasFakeCaret, isActive } = inputOTPContext.slots[index]
+  const { char, hasFakeCaret, isActive } = inputOTPContext.slots[index] ?? {}
 
   return (
     <div