+++ src/index.ts
@@ -4,7 +4,7 @@ export { cn } from "./lib/utils";
 export * from "./lib/formatters";
 export * from "./lib/export-utils";
 export * from "./lib/validators";
-export * from "./lib/indiaLocations";
+export * from "./lib/locations";
 export * from "./lib/blob-storage";
 export * from "./lib/masking";
 export * from "./lib/telemetry";